chore: add k2-components with light theme
