+++ package.json
@@ -14,6 +14,7 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
+    "@avalabs/k2-components": "4.12.0",
     "@docusaurus/core": "2.4.1",
     "@docusaurus/plugin-google-gtag": "2.4.1",
     "@docusaurus/preset-classic": "2.4.1",
@@ -27,6 +28,7 @@
     "prism-react-renderer": "1.3.5",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-icons": "4.10.1",
     "react-iframe": "1.8.5",
     "rehype-katex": "5",
     "remark-math": "3",
