+++ src/pages/Home/home.jsx
@@ -1,0 +1,91 @@
+import React from "react";
+import Layout from "@theme/Layout";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import styles from "./home.module.css";
+import Card from "../../components/Card";
+import { Typography } from "@avalabs/k2-components";
+
+export function Home() {
+  const context = useDocusaurusContext();
+  const { siteConfig = {} } = context;
+
+  return (
+    <main>
+      <br />
+      <h1 align="center" style={{ fontWeight: "750" }}>
+        Welcome to Avalanche Dev Docs
+      </h1>
+      <section className={styles.features}>
+        <div className="container">
+          <div className="row cards__container">
+            <Card
+              to="/build/subnet/hello-subnet"
+              header={{
+                label: "🚀 Launch Your First Subnet",
+              }}
+              body={{
+                label:
+                  "Start your Subnet development journey by creating a subnet in under five minutes",
+              }}
+            />
+
+            <Card
+              to="intro"
+              header={{
+                label: "🔺 Learn about Avalanche",
+              }}
+              body={{
+                label:
+                  "Discover how Subnets and Avalanche Consensus are revolutionizing Web3",
+              }}
+            />
+
+            <Card
+              to="/nodes/validate/add-a-validator"
+              header={{
+                label: "😎 Become a Validator",
+              }}
+              body={{
+                label:
+                  "Join Avalanche's Proof-of-Stake protocol to help secure the network and earn rewards",
+              }}
+            />
+
+            <Card
+              to="reference"
+              header={{
+                label: "💻 View Avalanche APIs",
+              }}
+              body={{
+                label:
+                  "Access low-level protocol interfaces to build your custom dapp",
+              }}
+            />
+
+            <Card
+              to="/build/dapp/launch-dapp"
+              header={{
+                label: "🛠️ Launch Your Dapp on Avalanche",
+              }}
+              body={{
+                label:
+                  "Learn everything you need to deploy an EVM-compatible smart contract",
+              }}
+            />
+
+            <Card
+              to="https://github.com/ava-labs/hypersdk"
+              header={{
+                label: "🛠️ HyperSDK",
+              }}
+              body={{
+                label:
+                  "Opinionated Framework for Building Hyper-Scalable Blockchains on Avalanche",
+              }}
+            />
+          </div>
+        </div>
+      </section>
+    </main>
+  );
+}
